Optimize some code readability

+++ PhiFansChart.cs
@@ -46,7 +46,7 @@ public static class PhiFansObject
     {
         public int type = 1;
         public int[] beat = new int[3];
-        public int positionX;
+        public float positionX;
         public float speed;
         public bool isAbove = true;
         public int[] holdEndBeat = new int[3];
